requirements.txt
svglib
cairo
reportlab
pillow
pycairo
rlPyCairo